fix circular dep issue
